requirements.txt
PyStaticConfiguration==0.9.0
PyYAML==3.11
six==1.10.0
tweepy==1.11
